+++ changelog.md
@@ -15,6 +15,12 @@ Preview features are treated as a separate branch and will not be included in th
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### <a name="unreleased"/> Unreleased
+
+#### Fixed
+
+- [NNNN](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/NNNN) ChangeFeedProcessor: Fixes AllVersionsAndDeletes cold-start regression introduced by #5617
+
 ### <a name="3.60.0-preview.0"/> [3.60.0-preview.0](https://www.nuget.org/packages/Microsoft.Azure.Cosmos/3.60.0-preview.0) - 2026-4-24
 
 #### Added